Fixed a few sharding bugs, sharding replication and pivot_point syncing
Overloaded the _in_sync function in the container/replicator to
return False if there are pivot_points becuase we really need to
make sure these are synced between the sharded containers.

Fixed a sharder replication problem where replicating the current
broker was deleting it locally becuause the wrong partition was being
passed to the replicate method. Making it think it was a handoff node.