Send a container PUT for new shard containers
To split a new container the sharder creates a new broker (db) and
then uses replication to send it out. This however wont cause the
hidden account to be updated with the new containers.
The 'lt' containers get added becuase we are sending a container PUT
to make sure the hidden account exists. This change sends container
a 'gt' container PUT at the same time to make sure the account and both
containers exist.