This patch has other fixes to the sharding_range branch
 - Improves the PivotRange.overlaps() method
 - container obj DELETE is pivot request aware
 - When sharding the parent is updated to have 0, 0 stats and/or
   deleted.
 - Other fixes

I've changed the way the root container's pivot_ranges gets updated
in the sharder. It now uses contaienr obj pivot PUT/DELETE requests,
but still not sure this is the best.
The reason these requests now exist is so pivot_ranges in the root can
get the sharded folders stats (object_count and bytes_used). And thought
maybe we should dogfood these requests and use them to update (add/delete)
the root contiainer. The old way (and the method is still in the code base)
is to use replication to update these, by creating a new broker, filling it
with the new items and the REPLICATE. The old method is more expensive but
bypasses the pending file.. which might be worth the extra effort.

Change-Id: Idff82e5aa2183f66d81f806306f67c0c6b3b55b2