Add substr parameter to the container-server
This patch adds a substr parameter to container listing:

  GET /v1/a/cont?substr=bob

This will return objects that contain the specific substring.
NOTE: Initally this is just some code to test out the possiblity, if
it goes forward the api (substr) may change. Also I still need to
update docs and add unit tests.

Change-Id: Idfafdc15624916ced61d73713fa89a9c7f43ced7