Add process level concurrency to container sync
This change introduces a new config option 'processes' to the
container sync daemon. The daemon spawns a new process per each
container to sync, up to the 'processes' limit. When the limit
is reached, the daemon waits for one of the processes to finish.

The 'spawning pattern' follows that of e.g. the container updater
The daemon process communicates with pipes with the child processes
to keep track of sync statistics. This, again, follows the container
updater pid2filename usage only with pipes instead of files.

The default for processes is set to 1, so that by default a single
child process is spawned in addition to the daemon.

This patch is closely related to [1] which adds thread level
concurrency to the sync daemon. [2] gives full details on
performance measurements showing the gain in throughput when
applying different levels of concurrency to the sync daemon.

[1] https://review.openstack.org/#/c/225338/
[2] https://ibm.box.com/s/oerppuv1fj907l2oysw0p39u3q1quvxi

Co-Authored-By: Oshrit Feder <oshritf@il.ibm.com>
Co-Authored-By: Alistair Coles <alistair.coles@hpe.com>

Change-Id: I8566b8b0d5511e93f8f5ea2eb2a25d0d3631bae8
Partial-Bug: #1068426