Stopped using OFFSET when looking for pivot
Turns out on _very_ large containers OFFSET is expensive
instead am using a new good enough version of get_possible_pivot
that uses a special optimisation inside SQLite's index structure.

SQLite Indexes are double-linked lists, so if you are using a SELECT
that hits an index and ask for just min or max, then it'll jump
directly to the bottom or top of the list without a table or index scan.

The new approach uses this optimisation, sends 2 SQL queries, one for
the min(name) and one for the max(name). We do our best guess and pick
a name to pivot on which happens to be the middle string between the 2
names. There is a new function in the container backend called
middle_str which finds the middle string.

The old OFFSET version would take a very long time on a container with
700M objects. Whereas the fuzzy approach on the same container took
almost nothing.

Change-Id: I0595232e91300dd866067d1fb7b736ddbb0e50a6