Added a sharding lock and other code cleanups
When sharding a very large container, the db_replicators will
go in and start replicating the new shard containers before
they have finished filling. This isn't too bad, except when
this triggers sharding shrinks on the new containers, then
confustion ensues, plus it means the sharded containers start
to using usync replication, which is less efficient then splitting
large containers.

The sharding lock is a .sharding file that is dropped in the parent
directory of the container database, and is removed when the sharding
has finished. This will trigger a rsync_then_merge which is more effient
with larger data.

Change-Id: I278225701e211038e60651fe2e7c2433639bff97