Started fixing up replication some more.
Now we always pass a filenames. (but is backwards compatible).
If it's rsync_then_merge, and the local and remote file is different,
we are using the new filename and unlinking the old.. but now we have
to be careful when we are pushing.

The idea will be we need to pass the db_state, and not push unsharded
over sharded.. this is still to do.

Change-Id: I51fff5042d9cec2c15f350ad0c3bae61c9a1ce4a