More replication fixes
Added a _can_push(info, rinfo) to db_replicator that
can be overridden by the db replicators (and is by the
container).

This method will tell the replicator if it can use rsync_then_merge
or if it'll have to fall back to usync.

In the case of sharding, we can't put and unsharded db over a
sharded one. Also we don't want to rsync_then_merge when one
of the containers are sharding (though they should be in an ideal
world.

Change-Id: I118c9fb6cd594f070e077cdf314aa23acb6866e7