Optionally log statsd metrics to swift recon cache
Swift logs a bunch of extra metrics to statsd, but only if you point it
to a statsd host. This patch allows you to add `log_stats` to your
configuration. When you do, a 1, 5 and 15 minute history of statsd
metrics are added to the recon cache.
These new recon files end in .stats, e.g:

  $ ls -1 /var/cache/swift/*.stats
  /var/cache/swift/object-auditor.stats
  /var/cache/swift/object-reconstructor.stats
  /var/cache/swift/object-replicator.stats
  /var/cache/swift/object-server.stats

A rough new CLI has been added to display this data. Currently
`swift-stats` will only gather and display the .stats on a given
server. The thought is if you want to gather data everywhere, we
could add a recon parameter.

Currently `swift-stats` takes in either a config path or a path
that represents the path to the recon_cache. If nothing is specified
it just assumes the recon_cache is '/var/cache/swift/'.
It will find all the .stat files, order them my account, container,
object and other (proxy) and then display them in a more readable
format.

Because statsd metrics can really be anything, it doesn't keep track
of known metrics, instead it just gives a 1, 5, 15 minute interval of
the triggered metrics.
Something like: http://paste.openstack.org/show/626132/

I'm thinking of maybe adding clays handoff counting script as an
optional cli parameter, so the tool could also be used to report that
at a glance.

The idea is, you can jump on a node and run swift-stats to see what
it's been up to. A bit like load avg for swift.

Change-Id: I71e41c2007e0f40af7eb624b402f3a8246f1b0bf