Add use_dynamic_reseller_names param to keystoneauth
If your using auth version 3 and want to have an account based of the
name of the project, then you can enable ``use_dynamic_reseller_name``
in conjuction with ``use_dynamic_reseller`` to do so. The form used
for the account would be::

    <reseller_prefix><project_domain_name>_<project_name>

Given a reseller_preix of 'AUTH_' as project_domain_name of 'Domain' and
a project name of 'Demo', this would generate::

    AUTH_Domain_Demo

NOTE: however, that if you ever change the domain or project name then
users accounts will change, loosing access to their old data, and
a large serverside copy would need to happen.
This feature however is really useful for federated environments however.

Change-Id: Ifad4526d09a4d29389594562fd7ca994125cdc37