Added a shardrange_graph_viewer cli script
To correctly use this script you need to have python-graphviz and
graphvix installed. If you do you can use it to generate a shard range
graph.

As it currently stands it only supports getting information from STDIN.
This was a choice because we're thinking of wrapping it in a docker
container to include all dependencies, so would make it easier to use.

Current usage is:

 $ python swift/cli/shardrange_graph_viewer.py --help
 usage: shardrange_graph_viewer.py [-h] [-i] [-f {svg,png,pdf}] [graph_name]

 Display ShardRange graphs

 positional arguments:
   graph_name            Name of the graph

 optional arguments:
   -h, --help            show this help message and exit
   -i, --info            Data is from container-info
   -f {svg,png,pdf}, --format {svg,png,pdf}
                         Format of the graph to produce

If you pass the '-i' it'll assume in the input is the output of
container-info:

  $ swift-container-info ... | python shardrange_graph_viewer.py -i

If you don't it'll assume json:

  $ swift-manage-shard-ranges .. show | python \
    shardrange_graph_viewer.py

Of course you can also load in some json from disk:

  $ python shardrange-graph-viewer.py < <path_to_json>

This will attempt to open the file if you can, but in anycase it'll drop
the graph file in whatever format you specify in the same folder you ran
the script from.

Then the analyse command lands, I plan to extend it to add colours for
the correct/selected path.

Change-Id: I5b02785631887f1a41430013ca39d35f5523baf3