request tracing: Add Otel request tracing support (decorator)
We've been wanting some micro benchmarks inside our WSGI environments,
like the proxy server, in order to test and find bottle necks etc.  But
on top of that, it'll also be awesome if we could support handing out a
trac_sig/key to a user and if they add this to their request path extra
tracing data could be dumped.

Initially we see a few ways of activating a trace:
 1. provide trace_sig which will be a HMAC from some trace key and expiry time.
 2. Trace every x request.
 3. Continue a trace if it finds it previously activated and trace
    context found inserted into a otel http header.

When one of these are met, a tracer is created and added to the request
environment which is used to trace the rest of the request.

Tracing is handled by a RequestTraceMiddleware wsgi mddleware.
This middleware is used to enable request tracing.
The request will start and stop tracing as the request hits this
middleware. As such, to get as much info as possible it's best to have it
as close to the front of the pipeline as possible.

The proxy has been instrumented so when making calls to storage nodes
the tracing context is injected into the sub requests allowing up to
trace a single request throughout the cluster. So long as the trace
middleware has also been included inside the storage node pipelines too.

Request tracing utilises the OpenTelemetry framework, which has now
become a swift dependency. We could pull this out and warn.

Because a trace is scoped by the request. There is a trace_data object that
is used to store some tracing metadata. Namely:

 - The call stack (so we can name middleware uniquely)
 - The current tracer object. As it's this object that creates spans.

Tracing is handled by a new class decorator, wsgi_trace, that can be added to a
middleware class. It'll use it's own `__call__` method that
automatically sets up a span and the tracing env for the middleware then
calls the wrapped class `__call__` function.

All the tracing code is in the swift.common.tracing namespace. And there
you'll find all the helper methods to further instrument code. To just
name a few:

  - trace_add: add a key value pair to the span.
  - new_trace_span: a context manager to help create new spans
  - trace_exception: attach a thrown exception to the span.

All the helper methods require passing in a wsgi env, this is where the
tracing object will live if it exists. If tracing ins't enabled for the
request, it's just a noop.

Unfortunately, tracing and eventlet don't always play nice. As such we
can't just use the global nature of the python otel implementation and
have to do some interesting scaffolding to get them to work together.
Such as, make the GreenAsyncPool trace aware, and also created
sub-classed GreenPile to TraceAwareGreenPile, so it can pass a span (if
it exists) into the eventlet environment, allowing them to be traced
correctly.
For example, sometimes an eventlet thread can loose the current context
so we need to force it.

Because SLOs are a heavily used middleware, we've heavily instrumented
SLOs, including the SegmentIterater response iterator
used by SLO. By the time the iter is being called by the client the
rest of the request spans have returned, so here is another time we need
to take our saved current span and force it back so the tracers can link
up correctly at the collection point, rather then starting "new" traces.

Signed-off-by: Matthew Oliver <matt@oliver.net.au>
Change-Id: I6c3765d43e0521dc0648ccd0956b93306d31e062